+++ Code/HuxEngine/src/de/hux/graphics/primitives/Vector2D.java
@@ -2,15 +2,15 @@ package de.hux.graphics.primitives;
 
 public class Vector2D
 {
-	public final int x;
-	public final int y;
+	public final float x;
+	public final float y;
 	
 	public Vector2D()
 	{
 		this(0, 0);
 	}
 	
-	public Vector2D(int x, int y)
+	public Vector2D(float x, float y)
 	{
 		this.x = x;
 		this.y = y;
